+++ packages/sdk/src/sdk-blockchains/flow/burn.test.ts
@@ -11,10 +11,10 @@ import { FlowBurn } from "./burn"
 describe("Flow burn", () => {
 	const { authUser1 } = createTestFlowAuth(fcl)
 	const wallet = new FlowWallet(fcl)
-	const sdk = createFlowSdk(wallet.fcl, "testnet", authUser1)
+	const sdk = createFlowSdk(wallet.fcl, "testnet", {}, authUser1)
 	const apis = createApisSdk("staging")
-	const mint = new FlowMint(sdk, apis)
-	const burn = new FlowBurn(sdk)
+	const mint = new FlowMint(sdk, apis, "testnet")
+	const burn = new FlowBurn(sdk, "testnet")
 
 	test.skip("Should burn NFT", async () => {
 		const itemId = await createTestItem(mint)
@@ -23,7 +23,7 @@ describe("Flow burn", () => {
 		const prepare = await burn.burn({ itemId })
 		const tx = await prepare.submit()
 		if (tx) {
-		  expect(tx.transaction.status).toEqual(4)
+			expect(tx.transaction.status).toEqual(4)
 		}
 	})
 })