+++ packages/sdk/src/sdk-blockchains/flow/buy.test.ts
@@ -13,11 +13,13 @@ import { FlowBuy } from "./buy"
 describe("Flow buy", () => {
 	const { authUser1 } = createTestFlowAuth(fcl)
 	const wallet = new FlowWallet(fcl)
-	const sdk = createFlowSdk(wallet.fcl, "testnet", authUser1)
+	const sdk = createFlowSdk(
+		wallet.fcl, "testnet", { basePath: "https://flow-api-staging.rarible.com" }, authUser1,
+	)
 	const apis = createApisSdk("staging")
-	const mint = new FlowMint(sdk, apis)
+	const mint = new FlowMint(sdk, apis, "testnet")
 	const sell = new FlowSell(sdk, apis)
-	const fill = new FlowBuy(sdk, apis)
+	const fill = new FlowBuy(sdk, apis, "testnet")
 
 	test.skip("Should buy flow NFT item", async () => {
 		const itemId = await createTestItem(mint)