+++ packages/sdk/src/sdk-blockchains/tezos/common/index.ts
@@ -1,7 +1,13 @@
 import type { ItemId, UnionAddress, Order, AssetType } from "@rarible/api-client"
 import { Blockchain } from "@rarible/api-client"
 // eslint-disable-next-line camelcase
-import type { Provider, TezosProvider, AssetType as TezosAssetType, Asset as TezosLibAsset } from "tezos-sdk-module/dist/common/base"
+import type {
+	Provider,
+	TezosProvider,
+	AssetType as TezosAssetType,
+	Asset as TezosLibAsset,
+	TezosNetwork,
+} from "tezos-sdk-module/dist/common/base"
 // eslint-disable-next-line camelcase
 import { get_public_key } from "tezos-sdk-module/dist/common/base"
 // eslint-disable-next-line camelcase
@@ -27,7 +33,6 @@ import type { Payout } from "@rarible/api-client/build/models/Payout"
 import type { Config } from "tezos-sdk-module"
 import type { UnionPart } from "../../../types/order/common"
 import type { CurrencyType } from "../../../common/domain"
-import type { TezosNetwork } from "../domain"
 
 export interface ITezosAPI {
 	collection: NftCollectionControllerApi,
@@ -98,6 +103,9 @@ export function getTezosBasePath(network: TezosNetwork): string {
 		case "hangzhou": {
 			return "https://rarible-api.functori.com"
 		}
+		case "mainnet": {
+			return "https://rarible-api-mainnet.functori.com"
+		}
 		default: {
 			throw new Error("Unsupported tezos network")
 		}
@@ -119,6 +127,19 @@ export function getMaybeTezosProvider(
 				config: {
 					exchange: "KT1ULGjK8FtaJ9QqCgJVN14B6tY76Ykaz6M8",
 					transfer_proxy: "KT1Qypf9A7DHoAeesu5hj8v6iKwHsJb1RUR2",
+					fees: new BigNumber(0),
+					nft_public: "",
+					mt_public: "",
+				},
+			}
+		}
+		case "mainnet": {
+			return {
+				tezos: provider,
+				api: `${getTezosBasePath(network)}/v0.1`,
+				config: {
+					exchange: "KT198mqFKkiWerXLmMCw69YB1i6yzYtmGVrC",
+					transfer_proxy: "KT1N2oby9tYmv5tjkGD1KyVzkDRCmgDkXgSD",
 					fees: new BigNumber(0),
 					nft_public: "",
 					mt_public: "",