+++ packages/sdk/src/sdk-blockchains/tezos/index.ts
@@ -1,11 +1,11 @@
 import type { TezosWallet } from "@rarible/sdk-wallet"
 import type { Maybe } from "@rarible/types/build/maybe"
+import type { TezosNetwork } from "tezos-sdk-module/dist/common/base"
 import type { IApisSdk, IRaribleInternalSdk } from "../../domain"
 import { TezosSell } from "./sell"
 import { TezosFill } from "./fill"
 import { TezosBid } from "./bid"
 import { getMaybeTezosProvider, getTezosAPIs } from "./common"
-import type { TezosNetwork } from "./domain"
 import { TezosMint } from "./mint"
 import { TezosTransfer } from "./transfer"
 import { TezosBurn } from "./burn"
@@ -23,27 +23,28 @@ export function createTezosSdk(
 	const apis = getTezosAPIs(network)
 	const maybeProvider = getMaybeTezosProvider(wallet?.provider, network)
 	const sellService = new TezosSell(maybeProvider, apis)
-	const mintService = new TezosMint(maybeProvider, apis)
+	const mintService = new TezosMint(maybeProvider, apis, network)
 	const bidService = new TezosBid(maybeProvider, apis)
+	const fillService = new TezosFill(maybeProvider, apis, network)
 
 	return {
 		nft: {
 			mint: mintService.mint,
-			burn: new TezosBurn(maybeProvider, apis).burn,
-			transfer: new TezosTransfer(maybeProvider, apis).transfer,
+			burn: new TezosBurn(maybeProvider, apis, network).burn,
+			transfer: new TezosTransfer(maybeProvider, apis, network).transfer,
 			generateTokenId: new TezosTokenId(maybeProvider, apis).generateTokenId,
-			deploy: new TezosDeploy(maybeProvider, apis).deployToken,
+			deploy: new TezosDeploy(maybeProvider, network).deployToken,
 			preprocessMeta: mintService.preprocessMeta,
 		},
 		order: {
-			fill: new TezosFill(maybeProvider, apis).fill,
-			buy: new TezosFill(maybeProvider, apis).fill,
-			acceptBid: new TezosFill(maybeProvider, apis).fill,
+			fill: fillService.fill,
+			buy: fillService.fill,
+			acceptBid: fillService.fill,
 			sell: sellService.sell,
 			sellUpdate: sellService.update,
 			bid: bidService.bid,
 			bidUpdate: bidService.update,
-			cancel: new TezosCancel(maybeProvider, apis).cancel,
+			cancel: new TezosCancel(maybeProvider, apis, network).cancel,
 		},
 		balances: {
 			getBalance: new TezosBalance(maybeProvider, apis).getBalance,