+++ CHANGELOG.md
@@ -3,6 +3,48 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.33](https://github.com/rarible/sdk/compare/v0.7.32...v0.7.33) (2021-12-15)
+
+
+### Bug Fixes
+
+* update flow-sdk ([#60](https://github.com/rarible/sdk/issues/60)) ([31ed68e](https://github.com/rarible/sdk/commit/31ed68e7f964a8098fc912bf7c4a1caa61039390))
+
+
+
+
+
+## [0.7.32](https://github.com/rarible/sdk/compare/v0.7.31...v0.7.32) (2021-12-14)
+
+
+### Bug Fixes
+
+* tezos up version ([ab37acb](https://github.com/rarible/sdk/commit/ab37acb87bda19d4e9a7b4022aaf53d94cd78503))
+* up yarn.lock file ([6800aad](https://github.com/rarible/sdk/commit/6800aadc71981e4d08369fed45a70758710b945a))
+
+
+
+
+
+## [0.7.31](https://github.com/rarible/sdk/compare/v0.7.30...v0.7.31) (2021-12-14)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [0.7.30](https://github.com/rarible/sdk/compare/v0.7.29...v0.7.30) (2021-12-13)
+
+
+### Bug Fixes
+
+* added royalties to preprocess metadata request ([327d546](https://github.com/rarible/sdk/commit/327d546c8234950999d678b2a34a1a7404491bc0))
+
+
+
+
+
 ## [0.7.29](https://github.com/rarible/sdk/compare/v0.7.28...v0.7.29) (2021-12-13)